preventDefault and additional event obj parameters now implemented
On the event object, there's now a preventDefault method that, when called,
will stop the hash from changing. Additionally, both the old and new hash are
exposed on the event object as .oldPath and .newPath simultaneously.

The choice of "Path" vs "Hash" in the name is conscientious, as this is typically
meant to represent a path and not just the hash. It also means the naming will
make more sense if/when HTML5 history is in play.